channel mapping: add ma_channel_mix_mode_matched
The ma_channel_mix_mode_simple does not necessarily map to the same
channel types you'd expect. For example, if you have a 6-channel input
and 8-channel output, you would get something like this with simple:

  Correct:  CHANNEL_FRONT_LEFT -> CHANNEL_FRONT_LEFT
  Correct:  CHANNEL_FRONT_RIGHT -> CHANNEL_FRONT_RIGHT
  Correct:  CHANNEL_FRONT_CENTER -> CHANNEL_FRONT_CENTER
  Correct:  CHANNEL_LFE -> CHANNEL_LFE
  Wrong:    CHANNEL_SIDE_LEFT -> CHANNEL_BACK_LEFT
  Wrong:    CHANNEL_SIDE_RIGHT -> CHANNEL_BACK_RIGHT

Instead of just assuming the input channel map is a truncation of the
output channel map and has channels in the same order, try to find the
matching channel type and set the weight for that.

Signed-off-by: Steven Noonan <steven@uplinklabs.net>